LPS-109189 Move IndexUpdaterUtil to portal-impl
